api/requirements.txt
CacheControl==0.12.11
cachetools==5.3.1
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==41.0.0
firebase-admin==6.1.0
Flask==2.2.5
google-api-core==2.11.0
google-api-python-client==2.88.0
google-auth==2.19.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpcio==1.54.2
grpcio-status==1.54.2
httplib2==0.22.0
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
msgpack==1.0.5
numpy==1.21.6
Pillow==9.5.0
proto-plus==1.22.2
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyparsing==3.0.9
requests==2.31.0
rsa==4.9
six==1.16.0
torch==1.13.1
typing_extensions==4.6.2
uritemplate==4.1.1
urllib3==1.26.16
Werkzeug==2.2.3
wincertstore==0.2
zipp==3.15.0
